* marshal.c (Init_marshal): new constant Marshal::MAJOR_VERSION   and Marshal::MINOR_VERSION.
* marshal.c (marshal_load): ruby_verbose test should be wrapped by
  RTEST().

* hash.c (rb_hash_index): should return nil (not the default
  value) if value is not in the hash.

* numeric.c (num_div): new method added.  alias to '/' which
  should be preserved even if '/' is redefined (e.g. by
  mathn). [new]

* bignum.c (rb_cstr2inum): "0 ff".hex should return 0, not 255.

* file.c (rb_file_s_expand_path): fixed using CharNext().


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@1658 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ lib/pstore.rb
@@ -79,7 +79,7 @@ class PStore
     throw :pstore_abort_transaction
   end
 
-  def transaction
+  def transaction(read_only=false)
     raise PStore::Error, "nested transaction" if @transaction
     begin
       @transaction = true
@@ -89,10 +89,13 @@ class PStore
 	file = File::open(@filename, "r+")
 	orig = true
       rescue Errno::ENOENT
+	raise if read_only
 	file = File::open(@filename, "w+")
       end
-      file.flock(File::LOCK_EX)
-      if orig
+      file.flock(read_only ? File::LOCK_SH : File::LOCK_EX)
+      if read_only
+	@table = Marshal::load(file)
+      elsif orig
 	content = file.read
 	@table = Marshal::load(content)
 	size = content.size
@@ -109,7 +112,7 @@ class PStore
 	@abort = true
 	raise
       ensure
-	unless @abort
+	if !read_only && !@abort
 	  file.rewind
 	  content = Marshal::dump(@table)
 	  if !md5 || size != content.size || md5 != MD5.new(content).digest
@@ -150,7 +153,7 @@ if __FILE__ == $0
     end
   end
 
-  db.transaction do
+  db.transaction(true) do
     p db["root"]
   end
 end
